Refactor jest providers to single file
