+++ src/components/MobilityPlatform/BicycleStands/components/BicycleStandContent/__tests__/BicycleStandContent.test.js
@@ -1,17 +1,8 @@
 // Link.react.test.js
 import React from 'react';
-import { ThemeProvider } from '@mui/material/styles';
-import { render } from '@testing-library/react';
-import { IntlProvider } from 'react-intl';
-import themes from '../../../../../../themes';
 import BicycleStandContent from '../index';
 import finnishTranslations from '../../../../../../i18n/fi';
-
-// Mock props for intl provider
-const intlMock = {
-  locale: 'fi',
-  messages: finnishTranslations,
-};
+import { getRenderWithProviders } from '../../../../../../../jestUtils';
 
 const mockProps = {
   bicycleStand: {
@@ -27,16 +18,7 @@ const mockProps = {
   },
 };
 
-// eslint-disable-next-line react/prop-types
-const Providers = ({ children }) => (
-  <IntlProvider {...intlMock}>
-    <ThemeProvider theme={themes.SMTheme}>
-      {children}
-    </ThemeProvider>
-  </IntlProvider>
-);
-
-const renderWithProviders = component => render(component, { wrapper: Providers });
+const renderWithProviders = getRenderWithProviders({});
 
 describe('<BicycleStandContent />', () => {
   it('should work', () => {
