+++ src/components/MobilityPlatform/Boating/Marinas/components/MarinasContent/__tests__/MarinasContent.test.js
@@ -1,17 +1,8 @@
 // Link.react.test.js
 import React from 'react';
-import { ThemeProvider } from '@mui/material/styles';
-import { render } from '@testing-library/react';
-import { IntlProvider } from 'react-intl';
-import themes from '../../../../../../../themes';
 import MarinasContent from '../index';
+import { getRenderWithProviders } from '../../../../../../../../jestUtils';
 import finnishTranslations from '../../../../../../../i18n/fi';
-
-// Mock props for intl provider
-const intlMock = {
-  locale: 'fi',
-  messages: finnishTranslations,
-};
 
 const mockProps = {
   name: 'Satama: Marina',
@@ -23,16 +14,7 @@ const mockProps = {
   }],
 };
 
-// eslint-disable-next-line react/prop-types
-const Providers = ({ children }) => (
-  <IntlProvider {...intlMock}>
-    <ThemeProvider theme={themes.SMTheme}>
-      {children}
-    </ThemeProvider>
-  </IntlProvider>
-);
-
-const renderWithProviders = component => render(component, { wrapper: Providers });
+const renderWithProviders = getRenderWithProviders({});
 
 describe('<MarinasContent />', () => {
   it('should work', () => {
