+++ src/views/MobilitySettingsView/components/FormLabel/__tests__/FormLabel.test.js
@@ -1,33 +1,15 @@
 // Link.react.test.js
 import React from 'react';
-import { ThemeProvider } from '@mui/material/styles';
-import { render } from '@testing-library/react';
-import { IntlProvider } from 'react-intl';
-import themes from '../../../../../themes';
 import FormLabel from '../index';
+import { getRenderWithProviders } from '../../../../../../jestUtils';
 import finnishTranslations from '../../../../../i18n/fi';
-
-// Mock props for intl provider
-const intlMock = {
-  locale: 'fi',
-  messages: finnishTranslations,
-};
 
 const mockProps = {
   msgId: 'mobilityPlatform.menu.showBicycleStands',
   checkedValue: false,
 };
 
-// eslint-disable-next-line react/prop-types
-const Providers = ({ children }) => (
-  <IntlProvider {...intlMock}>
-    <ThemeProvider theme={themes.SMTheme}>
-      {children}
-    </ThemeProvider>
-  </IntlProvider>
-);
-
-const renderWithProviders = component => render(component, { wrapper: Providers });
+const renderWithProviders = getRenderWithProviders({});
 
 describe('<FormLabel />', () => {
   it('should work', () => {
